\docType{package}
\name{estate-package}
\alias{estate}
\alias{estate-package}
\title{The estate Package}
\description{
  Real estate.
}
\details{
  \tabular{ll}{ Package: \tab estate\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2014-04-06\cr
  Depends: \tab R (>= 2.10.0)\cr Encoding: \tab UTF-8\cr
  License: \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr URL:
  \tab http://spark.rstudio.com\cr }

  Provides a function to scrape information from pap.fr.
}
\author{
  Bo Werth \email{bo.werth@gmail.com}
}
\keyword{package}

